+++ packages/spacecat-shared-google-client/CHANGELOG.md
@@ -1,3 +1,10 @@
+# [@adobe/spacecat-shared-google-client-v1.4.11](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.4.10...@adobe/spacecat-shared-google-client-v1.4.11) (2025-02-22)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#622](https://github.com/adobe/spacecat-shared/issues/622)) ([6552a61](https://github.com/adobe/spacecat-shared/commit/6552a61ebd8b83c0a1dec51d50d44cf3b97819c1))
+
 # [@adobe/spacecat-shared-google-client-v1.4.10](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.4.9...@adobe/spacecat-shared-google-client-v1.4.10) (2025-02-16)